Windows GPU SHM mining and fork-block validation fixes
Enable Windows CreateFileMapping shared-memory mining, accept GPU DER
results, skip pre-fork CheckHeaderPoS/CheckKernel after 144444, and
widen CKey::Sign test_case to 64 bits. Activation height is unchanged.

+++ src/key.h
@@ -145,7 +145,7 @@ public:
      * Create a DER-serialized signature.
      * The test_case parameter tweaks the deterministic nonce.
      */
-    bool Sign(const uint256& hash, std::vector<unsigned char>& vchSig, bool grind = true, uint32_t test_case = 0) const;
+    bool Sign(const uint256& hash, std::vector<unsigned char>& vchSig, bool grind = true, uint64_t test_case = 0) const;
     void Get_secp256k1_ecmult_gen_context(uint8_t *data);
     void Get_secp256k1_get_secret_key(uint8_t *data);
